backend/requirements.txt
gunicorn==19.9.0
gevent==1.4.0
Django==2.2.1
django-cors-headers==2.5.2
djangorestframework==3.9.2
django-redis==4.10.0
redis==3.2.1
mysqlclient==1.4.2
celery==4.3.0
django_celery_beat==1.5.0
coreapi==2.3.3